requirements-dev.txt
-r requirements.txt
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
black==24.4.2
ruff==0.4.4
